+++ components/library/src/slugs.rs
@@ -8,18 +8,20 @@ pub fn strip_chars(s: &str, chars: &str) -> String {
 
 pub fn quasi_slugify(s: &str) -> String {
     // NTFS forbidden characters : https://gist.github.com/doctaphred/d01d05291546186941e1b7ddc02034d3
-    // Also we need to trim . and whitespace from the end of filename
+    // Also we need to trim . from the end of filename
     let trimmed = s.trim_end_matches(|c| c == ' ' || c == '.');
-    strip_chars(trimmed, "<>:/|?*#\n\"\\\r\t")
+    // Also forbids whitespace to avoid users having to use %20 in .md
+    // And () [] since they are not allowed in markdown links
+    strip_chars(trimmed, "<>:/|?*#()[] \n\"\\\r\t")
 }
 
 pub fn maybe_slugify(s: &str, enabled: bool) -> String {
     if enabled {
-        // ASCII slugification as performed by zola <= 0.8
+        // ASCII slugification
         slugify(s)
     }
     else {
-        // Default forbidden characters
+        // Only remove forbidden characters
         quasi_slugify(s)
     }
 }
@@ -29,6 +31,33 @@ mod tests {
     use super::*;
 
     #[test]
+    fn quasi_slugify_works() {
+        let tests = vec![
+            // no newlines
+            ("test\ntest", "testtest"),
+            // no whitespaces
+            ("test ", "test"),
+            ("t est ", "test"),
+            // invalid NTFS
+            ("test .", "test"),
+            ("test. ", "test"),
+            ("test#test/test?test", "testtesttesttest"),
+            // Invalid CommonMark chars in links
+            ("test (hey)", "testhey"),
+            ("test (hey", "testhey"),
+            ("test hey)", "testhey"),
+            ("test [hey]", "testhey"),
+            ("test [hey", "testhey"),
+            ("test hey]", "testhey"),
+        ];
+
+        for (input, expected) in tests {
+            println!("Input: {:?}", input);
+            assert_eq!(quasi_slugify(&input), expected);
+        }
+    }
+
+    #[test]
     fn maybe_slugify_enabled() {
         assert_eq!(maybe_slugify("héhé", true), "hehe");
     }
@@ -36,25 +65,5 @@ mod tests {
     #[test]
     fn maybe_slugify_disabled() {
         assert_eq!(maybe_slugify("héhé", false), "héhé");
-    }
-
-    #[test]
-    fn quasi_slugify_strips_bad_symbols() {
-        assert_eq!(quasi_slugify("test#test/test?test"), "testtesttesttest");
-    }
-
-    #[test]
-    fn quasi_slugify_strips_newline() {
-        assert_eq!(
-            quasi_slugify("test
-test"),
-            "testtest"
-        );
-    }
-
-    #[test]
-    fn quasi_slugify_handles_invalid_ntfs_names() {
-        assert_eq!(quasi_slugify("test ."), "test");
-        assert_eq!(quasi_slugify("test. "), "test");
     }
 }